Optimized algorithm with manual tail recursion
The change was made possible by subsituting the recursive calls with a saved list of start times to be processed next.
This list is iterated over and "simulates" the next recursion batch.
The recursion itself is replaced with a label and jump, since PHP does not automatically optimize for tail recursion.